+++ src/components/EditSections/EditUserInfo/components/ProfilePicture/ProfilePicture.test.js
@@ -1,4 +1,8 @@
-import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
+import {
+  render,
+  screen,
+  fireEvent
+} from '@folio/jest-config-stripes/testing-library/react';
 import userEvent from '@folio/jest-config-stripes/testing-library/user-event';
 import profilePicData from 'fixtures/profilePicture';
 
@@ -11,15 +15,19 @@ jest.mock('../../../../../hooks', () => ({
   useProfilePicture: jest.fn(),
 }));
 
-const props = {
-  label: 'Profile picture',
-  profilePictureLink: 'profilePictureLink'
+const defaultProps = {
+  profilePictureId: 'https://folio.org/wp-content/uploads/2023/08/folio-site-general-Illustration-social-image-1200.jpg',
+  form: {
+    change: jest.fn(),
+  }
 };
+
+const renderProfilePicture = (props) => render(<ProfilePicture {...props} />);
 
 describe('Profile Picture', () => {
   beforeEach(() => {
     useProfilePicture.mockClear().mockReturnValue(profilePicData.profile_picture_blob);
-    render(<ProfilePicture {...props} />);
+    renderProfilePicture(defaultProps);
   });
 
   it('should display Profile picture', () => {
@@ -28,7 +36,7 @@ describe('Profile Picture', () => {
 
   it('Image to be displayed with correct src', () => {
     const image = screen.getByTestId('profile-picture');
-    expect(image.src).toContain('profilePictureLink');
+    expect(image.src).toContain('https://folio.org/wp-content/uploads/2023/08/folio-site-general-Illustration-social-image-1200.jpg');
   });
 
   it('Update button to be displayed', () => {
@@ -54,5 +62,28 @@ describe('Profile Picture', () => {
     await userEvent.click(updateButton);
 
     expect(screen.getByText('Icon (trash)')).toBeInTheDocument();
+  });
+
+  it('should render modal', async () => {
+    const updateButton = screen.getByTestId('updateProfilePictureDropdown');
+    await userEvent.click(updateButton);
+    const externalURLButton = screen.getByTestId('externalURL');
+    await userEvent.click(externalURLButton);
+
+    expect(screen.getByText('ui-users.information.profilePicture.externalLink.modal.externalURL')).toBeInTheDocument();
+  });
+
+  it('should call save handler', async () => {
+    const updateButton = screen.getByTestId('updateProfilePictureDropdown');
+    await userEvent.click(updateButton);
+    const externalURLButton = screen.getByTestId('externalURL');
+    await userEvent.click(externalURLButton);
+    const saveButton = screen.getByRole('button', { name: 'ui-users.save' });
+    const inputElement = screen.getByLabelText('ui-users.information.profilePicture.externalLink.modal.externalURL');
+
+    fireEvent.change(inputElement, { target: { value: 'https://upload.wikimedia.org/wikipedia/commons/e/e2/FOLIO_400x400.jpg' } });
+    await userEvent.click(saveButton);
+    const image = screen.getByTestId('profile-picture');
+    expect(expect(image.src).toContain('https://upload.wikimedia.org/wikipedia/commons/e/e2/FOLIO_400x400.jpg'));
   });
 });