+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * User Information in User Edit to display profile picture and update button set. Refs UIU-3005.
 * Update request header for pay several Fees/fines. Refs UIU-3040.
 * Use correct field for display remaining in list. Refs UIU-3049.
+* Link externally hosted image as user profile picture. Refs UIU-2975.
 
 ## [10.0.4](https://github.com/folio-org/ui-users/tree/v10.0.4) (2023-11-10)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v10.0.3...v10.0.4)